consistent namespacing, remove unnecessary requires
